finished restore data features

+++ app/Http/Controllers/ProductController.php
@@ -119,6 +119,7 @@ class ProductController extends Controller
      */
     public function indexDeleted()
     {
+        Gate::authorize('admin');
 
         $deletedProducts = Product::onlyTrashed()->get();
 
@@ -130,6 +131,8 @@ class ProductController extends Controller
      */
     public function showDeleted(string $id)
     {
+        Gate::authorize('admin');
+
         $product = Product::onlyTrashed()->findOrFail($id);
 
         return view('pages.products.deleted.show', compact('product'));
@@ -140,24 +143,33 @@ class ProductController extends Controller
      */
     public function fullyDelete(string $id)
     {
+        Gate::authorize('admin');
+
         Product::forceDestroy($id);
 
         return redirect('products/deleted')->with('success', 'Product permanently deleted.');
     }
 
     /**
-     * Restore deleted specific product from database
+     * Restore deleted specific product from database.
      */
     public function restoreDeleted(string $id)
     {
+        Gate::authorize('admin');
+
         $product = Product::onlyTrashed()->findOrFail($id);
         $product->restore();
 
         return redirect('products/deleted')->with('success', 'Product successfully restored');
     }
 
+    /**
+     * Delete All product in soft deletes.
+     */
     public function deleteAll()
     {
+        Gate::authorize('admin');
+        
         Product::onlyTrashed()->forceDelete();
         return redirect('products/deleted')->with('success', 'All products are successfully deleted');
     }
